crypto: Centralize function decoding
